+++ client/src/views/Widgets/Spotify.js
@@ -1,41 +1,9 @@
 import { useEffect, useState } from "react";
-import Axios from 'axios'
 
 import Login from "../LoginPage/LoginSpotify"
-import Cookies from "js-cookie";
-import SpotifyLastRelease from "../../components/SpotifyLastRelease";
 import dropDownCountryMenu from "../../components/dropDownCountryMenu";
-
-import Carousel from 'react-elastic-carousel';
-
-function getCountryLastRelease(country, setSpotifyLastRelease) {
-    var spotifyCookie = JSON.parse(Cookies.get("spotify"))
-
-    Axios.post("http://localhost:8080/spotify/release", {
-        token: spotifyCookie.access_token,
-        country: country
-    }).then((response) => {
-        if (response.data) {
-            setSpotifyLastRelease(JSON.parse(JSON.stringify(response.data)))
-        }
-    })
-}
-
-const mycarrousel = (spotifyLastReleaseItem) => {
-    if (!spotifyLastReleaseItem)
-        return (null)
-
-    const breakPoints = [
-        { width: 1, itemsToShow: 1 },
-    ];
-    const items = [0, 1, 2];
-
-    return (
-        <Carousel breakPoints={breakPoints}>
-            {items.map((item) => (SpotifyLastRelease(spotifyLastReleaseItem.albums.items[item])))}
-        </Carousel>
-    )
-}
+import getCountryLastRelease from "../../components/getCountryLastRelease"
+import releaseCarousel from "../../components/releaseCarousel"
 
 export default function Spotify() {
     const [spotifyConnected, setSpotifyConnected] = useState(false);
@@ -58,7 +26,7 @@ export default function Spotify() {
         <>
             <center>
                 {dropDownCountryMenu(anchorEl, setAnchorEl, setCountry)}
-                {mycarrousel(spotifyLastRelease)}
+                {releaseCarousel(spotifyLastRelease)}
             </center>
         </>
     )
